* develop:
  MARS-529: Make ADSM datahandles 'moveable'
  MARS-519 work on param matching
  MARS-519 debug info
  MARS-519: check param in legacy import
  MARS-519: work on tools
  MARS-519: code tidy
  MARS-519: code tidy
  MARS-519: code tidy
  MARS-519: work on roots
  MARS-519: work on roots
  MARS-519: work on tools
  MARS-519 work on roots
  MARS-519 work on tools

+++ src/fdb5/toc/WipeVisitor.cc
@@ -25,11 +25,23 @@ namespace fdb5 {
 //----------------------------------------------------------------------------------------------------------------------
 
 class StdDir {
+
     DIR *d_;
-public:
-    StdDir(const eckit::PathName& p) { d_ = opendir(p.asString().c_str());}
-    ~StdDir()                 { if(d_) closedir(d_);    }
-    operator DIR*()           { return d_;              }
+
+  public:
+
+    StdDir(const eckit::PathName &p) {
+        d_ = opendir(p.localPath());
+    }
+
+    ~StdDir() {
+        if (d_) closedir(d_);
+    }
+
+    operator DIR *() {
+        return d_;
+    }
+
 };
 
 
@@ -38,9 +50,17 @@ public:
 WipeVisitor::WipeVisitor(const eckit::PathName &directory) :
     ReportVisitor(directory) {
 
-    eckit::Log::info() << "Scanning " << directory_  << std::endl;
-    scan(directory_);
-    eckit::Log::info() << "Found " << eckit::Plural(files_.size(), "file") << ", size: " << eckit::Bytes(total_) << std::endl;
+    eckit::Log::info() << "Scanning "
+                       << directory_
+                       << std::endl;
+
+    scan(directory_.realName());
+
+    eckit::Log::info() << "Found "
+                       << eckit::Plural(files_.size(), "file")
+                       << ", size: "
+                       << eckit::Bytes(total_)
+                       << std::endl;
 
 }
 
@@ -50,56 +70,37 @@ void WipeVisitor::scan(const eckit::PathName &directory) {
 
     StdDir d(directory);
 
-    if(d == 0)
-    {
+    if (d == 0) {
         throw eckit::FailedSystemCall(std::string("opendir(") + std::string(directory) + ")");
     }
 
-    struct dirent buf;
+    struct dirent *e;
 
+    while ( (e = readdir(d)) != 0) {
 
-    for(;;)
-    {
-        struct dirent *e;
-#ifdef EC_HAVE_READDIR_R
-        errno = 0;
-        if(readdir_r(d, &buf, &e) != 0)
-        {
-            if(errno)
-                throw eckit::FailedSystemCall("readdir_r");
-            else
-                e = 0;
-        }
-#else
-        e = readdir(d);
-#endif
-
-        if(e == 0)
-            break;
-
-        if(e->d_name[0] == '.')
-            if(e->d_name[1] == 0 || (e->d_name[1] =='.' && e->d_name[2] == 0))
+        if (e->d_name[0] == '.') {
+            if (e->d_name[1] == 0 || (e->d_name[1] == '.' && e->d_name[2] == 0)) {
                 continue;
-
+            }
+        }
 
         eckit::PathName path(directory / e->d_name);
-        files_.insert( path.realName() );
-
+        files_.insert( path );
 
         eckit::Stat::Struct info;
-        SYSCALL(eckit::Stat::lstat(path.asString().c_str(), &info));
-        if(S_ISDIR(info.st_mode)) {
+        SYSCALL(eckit::Stat::lstat(path.localPath(), &info));
+
+        if (S_ISDIR(info.st_mode)) {
             scan(path);
-        }
-        else {
+        } else {
             total_ += path.size();
         }
     }
 }
 
-const eckit::PathName& WipeVisitor::mark(const eckit::PathName& path) const {
+const eckit::PathName &WipeVisitor::mark(const eckit::PathName &path) const {
     std::set<eckit::PathName>::iterator j = files_.find(path.realName());
-    if(j == files_.end()) {
+    if (j == files_.end()) {
         std::ostringstream oss;
         oss << "Unexpected file " << path << ". The file may have been added after the start of the command";
         throw eckit::UserError(oss.str());
@@ -129,7 +130,7 @@ void WipeVisitor::report(std::ostream &out) const {
             cnt++;
         }
     }
-    if(cnt == 0) {
+    if (cnt == 0) {
         out << "    - NONE -" << std::endl;
     }
     out << std::endl;
@@ -137,10 +138,10 @@ void WipeVisitor::report(std::ostream &out) const {
     cnt = 0;
     out << "Index files to be deleted:" << std::endl;
     for (std::map<eckit::PathName, size_t>::const_iterator i = indexUsage_.begin(); i != indexUsage_.end(); ++i) {
-            out << "    " << mark(i->first) << std::endl;
-            cnt++;
+        out << "    " << mark(i->first) << std::endl;
+        cnt++;
     }
-     if(cnt == 0) {
+    if (cnt == 0) {
         out << "    - NONE -" << std::endl;
     }
     out << std::endl;
@@ -148,10 +149,10 @@ void WipeVisitor::report(std::ostream &out) const {
     cnt = 0;
     out << "Other files:" << std::endl;
     for (std::set<eckit::PathName>::const_reverse_iterator i = files_.rbegin(); i != files_.rend(); ++i) {
-            out << "    " << (*i) << std::endl;
-            cnt++;
+        out << "    " << (*i) << std::endl;
+        cnt++;
     }
-    if(cnt == 0) {
+    if (cnt == 0) {
         out << "    - NONE -" << std::endl;
     }
     out << std::endl;
@@ -173,14 +174,13 @@ void WipeVisitor::wipe(std::ostream &out) const {
     }
 
     for (std::map<eckit::PathName, size_t>::const_iterator i = indexUsage_.begin(); i != indexUsage_.end(); ++i) {
-           i->first.unlink();
+        i->first.unlink();
     }
 
     for (std::set<eckit::PathName>::const_reverse_iterator i = files_.rbegin(); i != files_.rend(); ++i) {
-        if(i->isDir()) {
+        if (i->isDir() && !i->isLink()) {
             i->rmdir();
-        }
-        else {
+        } else {
             i->unlink();
         }
     }
